pyproject.toml
[build-system]
requires = ["flit_core >=3.9.0,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pyfunc_agent"
description = "Example of using agents to call python functions."
version = "0.0.0"
authors = [
    {name = "Hickmann, Kyle", email = "hickmank@gmail.com"},
]
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "langchain",
    "langgraph",
    "langchain_ollama",
    "pyyaml",
    "streamlit",
    "ruff",
    "sphinx",
    "coverage",
    "furo"
    ]
requires-python = ">=3.9"
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = ["Private :: Do Not Upload",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
]

[project.urls]
Source = "https://github.com/hickmank/pyFunc-Agent"

# This section instructs Flit to include *.json under prompts/ automatically.
[tool.flit.include]
"src/pyfunc_agent/prompts/*.yaml" = "prompts/"

[tool.ruff]
line-length = 89
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 89

[tool.ruff.lint]
select = [
    "E",
    "F",
    "D",
    "UP",
    "W",
    "ANN"]

ignore = ["ANN101"]  # Wants a `type` defined for `self` argument!

[tool.ruff.lint.pydocstyle]
convention = "google"